Unified interface to CAD exports
This commit unifies the interfaces to the following
file formats. BRep, IGES, and STEP.
This is done by using one common base class the defines
the interface. The private methods WriteImpl must be
implemented for the concrete writing procedure.
Also fixed some compiler warnings.

Addresses issue #122

+++ src/guide_curves/CCPACSGuideCurves.cpp
@@ -57,7 +57,7 @@ void CCPACSGuideCurves::ReadCPACS(TixiDocumentHandle tixiHandle, const std::stri
 
     std::vector<CCPACSGuideCurve*> curves;
     ReadContainerElement(tixiHandle, segmentXPath + "/guideCurves", "guideCurve", 1, curves);
-    for (int i = 0; i < curves.size(); i++) {
+    for (size_t i = 0; i < curves.size(); i++) {
         guideCurves[curves[i]->GetUID()] = PCCPACSGuideCurve(curves[i]);
     }
 }
